+++ .github/actions/loki/test/entrypoint.sh
@@ -1,9 +1,3 @@
 #!/bin/sh -ex
 
-cd "$INPUT_WORKING_DIR"
-yarn loki \
-  test \
-  --chromeFlags='--headless --no-sandbox --disable-gpu --disable-features=VizDisplayCompositor' \
-  --verboseRenderer \
-  --requireReference \
-  --reactUri file:$INPUT_REACT_URI
+yarn run loki:test-ci --concurrency=1