+++ packages/fuselage-ui-kit/src/contexts/kitContext.ts
@@ -1,4 +1,4 @@
-import { createContext } from 'react';
+import { createContext, useContext } from 'react';
 
 type ActionParams = {
   blockId: string;
@@ -12,8 +12,8 @@ type UiKitContext = {
   action: (state: ActionParams) => Promise<void> | void;
   state: (state: ActionParams) => Promise<void> | void;
   appId: string;
-  errors: Record<string, string>;
-  values: Record<string, { value: string }>;
+  errors?: Record<string, string>;
+  values: Record<string, { value: string } | undefined>;
   viewId?: string;
 };
 
@@ -26,3 +26,20 @@ export const defaultContext = {
 };
 
 export const kitContext = createContext<UiKitContext>(defaultContext);
+
+export const useUiKitContext = () => useContext(kitContext);
+
+export const useUiKitStateValue = <T extends string | number | undefined>(
+  actionId: string,
+  initialValue: T
+): {
+  value: T;
+  error: string | undefined;
+} => {
+  const { values, errors } = useUiKitContext();
+
+  return {
+    value: (values && (values[actionId]?.value as T)) ?? initialValue,
+    error: errors && errors[actionId],
+  };
+};