+++ packages/fuselage/src/components/Box/index.d.ts
@@ -167,3 +167,14 @@ export { default as AnimatedVisibility } from './AnimatedVisibility';
 export { default as Flex } from './Flex';
 export { default as Position, PositionAnimated } from './Position';
 export { default as Scrollable } from './Scrollable';
+
+export const useArrayLikeClassNameProp: <
+  T extends {
+    className?:
+      | string
+      | ReturnType<typeof css>
+      | (string | ReturnType<typeof css>)[];
+  }
+>(
+  props: T
+) => T & { className: string };