+++ packages/fuselage/src/components/Message/MessageMetrics/MessageMetricsItem/MessageMetricsItem.tsx
@@ -1,0 +1,5 @@
+import React, { FC, HTMLAttributes } from 'react';
+
+export const MessageMetricsItem: FC<HTMLAttributes<HTMLDivElement>> = (
+  props
+) => <div className='rcx-message-metrics__item' {...props} />;