+++ packages/fuselage/src/components/Message/MessageRoles.tsx
@@ -1,0 +1,10 @@
+import React, { FC } from 'react';
+
+export const MessageRoles: FC = function MessageRoles(props) {
+  return (
+    <div
+      className='rcx-box rcx-box--full rcx-message-header__roles'
+      {...props}
+    />
+  );
+};