+++ packages/fuselage/src/components/Message/Messages.stories.tsx
@@ -1,25 +1,14 @@
 import React from 'react';
 
-import { ThreadMessage } from '.';
+import Message from '.';
 import { Box, Avatar } from '..';
-import {
-  Message,
-  MessageBody,
-  MessageContainer,
-  MessageHeader,
-  MessageLeftContainer,
-  MessageName,
-  MessageRole,
-  MessageTimestamp,
-  MessageUsername,
-} from './Message';
 import { MessageDivider } from './MessageDivider';
-// import { MessageMetricsItem } from './MessageMetrics';
-import { MessageReactions } from './MessageReactions';
-import { MessageToolbox } from './MessageToolbox';
+import MessageReactions from './MessageReactions';
+import MessageToolbox from './MessageToolbox';
+import ThreadMessage from './ThreadMessage';
 
 export default {
-  title: 'Messages/Message',
+  title: 'Message/Message',
   component: Message,
   parameters: {
     jest: ['Message.spec.tsx'],
@@ -30,7 +19,7 @@ export const Default = () => (
   <Box>
     <MessageDivider>May, 24, 2020</MessageDivider>
     <Message className='customclass' clickable>
-      <MessageLeftContainer>
+      <Message.LeftContainer>
         <Avatar
           url='data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAgGBgcGBQgHBwcJCQgKDBQNDAsLDBkSEw8UHRofHh0aHBwgJC
                 4nICIsIxwcKDcpLDAxNDQ0Hyc5PTgyPC4zNDL/2wBDAQkJCQwLDBgNDRgyIRwhMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMj
@@ -45,31 +34,31 @@ export const Default = () => (
                 UH2oorkV10pRc7b1zXb/hZOzuJvM86QWEXeELxOzHSIPcmiiiunVlF2RNTpRkrs//Z'
           size={'x36'}
         />
-      </MessageLeftContainer>
-      <MessageContainer>
-        <MessageHeader>
-          <MessageName>Haylie George</MessageName>
-          <MessageUsername>@haylie.george</MessageUsername>
-          <MessageRole>Admin</MessageRole>
-          <MessageRole>User</MessageRole>
-          <MessageRole>Owner</MessageRole>
-          <MessageTimestamp>12:00 PM</MessageTimestamp>
-        </MessageHeader>
-        <MessageBody>
+      </Message.LeftContainer>
+      <Message.Container>
+        <Message.Header>
+          <Message.Name>Haylie George</Message.Name>
+          <Message.Username>@haylie.george</Message.Username>
+          <Message.Role>Admin</Message.Role>
+          <Message.Role>User</Message.Role>
+          <Message.Role>Owner</Message.Role>
+          <Message.Timestamp>12:00 PM</Message.Timestamp>
+        </Message.Header>
+        <Message.Body>
           Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
           nisi ut aliquip ex ea commodo consequat a duis aute irure dolor in
           reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
           pariatur. Consectetur adipiscing elit, sed do eiusmod tempor
           incididunt ut labore et dolore magna aliqua. Ut enim ad minim
           veniam...
-        </MessageBody>
+        </Message.Body>
         <MessageReactions>
-          <MessageReactions.Reaction counter={1} />
+          <MessageReactions.Reaction mine counter={1} />
           <MessageReactions.Reaction counter={2} />
           <MessageReactions.Reaction counter={3} />
           <MessageReactions.Action />
         </MessageReactions>
-      </MessageContainer>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -85,7 +74,7 @@ export const WithSequential = () => (
   <Box>
     <MessageDivider>May, 24, 2020</MessageDivider>
     <Message className='customclass' clickable>
-      <MessageLeftContainer>
+      <Message.LeftContainer>
         <Avatar
           url='data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAgGBgcGBQgHBwcJCQgKDBQNDAsLDBkSEw8UHRofHh0aHBwgJC
                 4nICIsIxwcKDcpLDAxNDQ0Hyc5PTgyPC4zNDL/2wBDAQkJCQwLDBgNDRgyIRwhMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMj
@@ -100,31 +89,31 @@ export const WithSequential = () => (
                 UH2oorkV10pRc7b1zXb/hZOzuJvM86QWEXeELxOzHSIPcmiiiunVlF2RNTpRkrs//Z'
           size={'x36'}
         />
-      </MessageLeftContainer>
-      <MessageContainer>
-        <MessageHeader>
-          <MessageName>Haylie George</MessageName>
-          <MessageUsername>@haylie.george</MessageUsername>
-          <MessageRole>Admin</MessageRole>
-          <MessageRole>User</MessageRole>
-          <MessageRole>Owner</MessageRole>
-          <MessageTimestamp>12:00 PM</MessageTimestamp>
-        </MessageHeader>
-        <MessageBody>
+      </Message.LeftContainer>
+      <Message.Container>
+        <Message.Header>
+          <Message.Name>Haylie George</Message.Name>
+          <Message.Username>@haylie.george</Message.Username>
+          <Message.Role>Admin</Message.Role>
+          <Message.Role>User</Message.Role>
+          <Message.Role>Owner</Message.Role>
+          <Message.Timestamp>12:00 PM</Message.Timestamp>
+        </Message.Header>
+        <Message.Body>
           Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
           nisi ut aliquip ex ea commodo consequat a duis aute irure dolor in
           reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
           pariatur. Consectetur adipiscing elit, sed do eiusmod tempor
           incididunt ut labore et dolore magna aliqua. Ut enim ad minim
           veniam...
-        </MessageBody>
+        </Message.Body>
         <MessageReactions>
           <MessageReactions.Reaction counter={1} />
           <MessageReactions.Reaction counter={2} />
           <MessageReactions.Reaction counter={3} />
           <MessageReactions.Action />
         </MessageReactions>
-      </MessageContainer>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -134,12 +123,12 @@ export const WithSequential = () => (
       </MessageToolbox.Wrapper>
     </Message>
     <Message className='customclass' clickable sequential>
-      <MessageLeftContainer></MessageLeftContainer>
-      <MessageContainer>
-        <MessageBody>
+      <Message.LeftContainer></Message.LeftContainer>
+      <Message.Container>
+        <Message.Body>
           Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua
-        </MessageBody>
-      </MessageContainer>
+        </Message.Body>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -149,20 +138,20 @@ export const WithSequential = () => (
       </MessageToolbox.Wrapper>
     </Message>
     <Message className='customclass' clickable sequential>
-      <MessageLeftContainer></MessageLeftContainer>
-      <MessageContainer>
-        <MessageBody>
+      <Message.LeftContainer></Message.LeftContainer>
+      <Message.Container>
+        <Message.Body>
           Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua
-        </MessageBody>
-      </MessageContainer>
+        </Message.Body>
+      </Message.Container>
     </Message>
     <Message className='customclass' clickable sequential>
-      <MessageLeftContainer></MessageLeftContainer>
-      <MessageContainer>
-        <MessageBody>
+      <Message.LeftContainer></Message.LeftContainer>
+      <Message.Container>
+        <Message.Body>
           Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua
-        </MessageBody>
-      </MessageContainer>
+        </Message.Body>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -178,7 +167,7 @@ export const MessageWithThread = () => (
   <Box>
     <MessageDivider unreadLabel='Unread'>May, 24, 2020</MessageDivider>
     <Message className='customclass' clickable>
-      <MessageLeftContainer>
+      <Message.LeftContainer>
         <Avatar
           url='data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAgGBgcGBQgHBwcJCQgKDBQNDAsLDBkSEw8UHRofHh0aHBwgJC
                 4nICIsIxwcKDcpLDAxNDQ0Hyc5PTgyPC4zNDL/2wBDAQkJCQwLDBgNDRgyIRwhMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMj
@@ -193,31 +182,31 @@ export const MessageWithThread = () => (
                 UH2oorkV10pRc7b1zXb/hZOzuJvM86QWEXeELxOzHSIPcmiiiunVlF2RNTpRkrs//Z'
           size={'x36'}
         />
-      </MessageLeftContainer>
-      <MessageContainer>
-        <MessageHeader>
-          <MessageName>Haylie George</MessageName>
-          <MessageUsername>@haylie.george</MessageUsername>
-          <MessageRole>Admin</MessageRole>
-          <MessageRole>User</MessageRole>
-          <MessageRole>Owner</MessageRole>
-          <MessageTimestamp>12:00 PM</MessageTimestamp>
-        </MessageHeader>
-        <MessageBody>
+      </Message.LeftContainer>
+      <Message.Container>
+        <Message.Header>
+          <Message.Name>Haylie George</Message.Name>
+          <Message.Username>@haylie.george</Message.Username>
+          <Message.Role>Admin</Message.Role>
+          <Message.Role>User</Message.Role>
+          <Message.Role>Owner</Message.Role>
+          <Message.Timestamp>12:00 PM</Message.Timestamp>
+        </Message.Header>
+        <Message.Body>
           Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
           nisi ut aliquip ex ea commodo consequat a duis aute irure dolor in
           reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
           pariatur. Consectetur adipiscing elit, sed do eiusmod tempor
           incididunt ut labore et dolore magna aliqua. Ut enim ad minim
           veniam...
-        </MessageBody>
+        </Message.Body>
         <MessageReactions>
           <MessageReactions.Reaction counter={1} />
           <MessageReactions.Reaction counter={2} />
           <MessageReactions.Reaction counter={3} />
           <MessageReactions.Action />
         </MessageReactions>
-      </MessageContainer>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -227,12 +216,12 @@ export const MessageWithThread = () => (
       </MessageToolbox.Wrapper>
     </Message>
     <Message className='customclass' clickable sequential>
-      <MessageLeftContainer></MessageLeftContainer>
-      <MessageContainer>
-        <MessageBody>
+      <Message.LeftContainer></Message.LeftContainer>
+      <Message.Container>
+        <Message.Body>
           Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua
-        </MessageBody>
-      </MessageContainer>
+        </Message.Body>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -242,20 +231,20 @@ export const MessageWithThread = () => (
       </MessageToolbox.Wrapper>
     </Message>
     <Message className='customclass' clickable sequential>
-      <MessageLeftContainer></MessageLeftContainer>
-      <MessageContainer>
-        <MessageBody>
+      <Message.LeftContainer></Message.LeftContainer>
+      <Message.Container>
+        <Message.Body>
           Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua
-        </MessageBody>
-      </MessageContainer>
+        </Message.Body>
+      </Message.Container>
     </Message>
     <Message className='customclass' clickable sequential>
-      <MessageLeftContainer></MessageLeftContainer>
-      <MessageContainer>
-        <MessageBody>
+      <Message.LeftContainer></Message.LeftContainer>
+      <Message.Container>
+        <Message.Body>
           Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua
-        </MessageBody>
-      </MessageContainer>
+        </Message.Body>
+      </Message.Container>
       <MessageToolbox.Wrapper>
         <MessageToolbox>
           <MessageToolbox.Item icon='quote' />
@@ -295,7 +284,7 @@ export const MessageWithThread = () => (
                 6hY+0yM88TzeNZY4luYwpVYyduOfrvhPTnr0pXSX9y5mCsyJMdyxxvwq599em+taItqCSNc90ChvZRUruUcT0JiO18Elpk7t8v41LWzacxkB
                 SuvjQ/FFJayjDWrCTepAQ2vUH0oo/Jk3ovpwJJeVCP5CN+lFFaaMqy+nAyuChvrTI2kN9JAsi2ZOy4IBHMnkSCP+iqBexSWdxLazoUljJVlP
                 UH2oorkV10pRc7b1zXb/hZOzuJvM86QWEXeELxOzHSIPcmiiiunVlF2RNTpRkrs//Z'
-            size={'x16'}
+            size='x16'
           />
         </ThreadMessage.LeftContainer>
         <ThreadMessage.Container>