+++ packages/fuselage/src/components/Options/Option/OptionMenu.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 
 const OptionMenu: FC = (props) => (
-  <div className='rcx-box--animated rcx-option__menu-wraper' {...props} />
+  <div className='rcx-box--animated rcx-option__menu-wrapper' {...props} />
 );
 
 export default OptionMenu;