+++ packages/onboarding-ui/src/pages/ResetPasswordConfirmationPage/ResetPasswordConfirmationPage.spec.tsx
@@ -1,0 +1,9 @@
+import ReactDOM from 'react-dom';
+
+import ResetPasswordConfirmationPage from './ResetPasswordConfirmationPage';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<ResetPasswordConfirmationPage />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});