+++ packages/ui-kit/src/rendering/surfaces/UiKitParserMessage.ts
@@ -2,6 +2,7 @@ import { ActionsBlock } from '../../blocks/layout/ActionsBlock';
 import { ContextBlock } from '../../blocks/layout/ContextBlock';
 import { DividerBlock } from '../../blocks/layout/DividerBlock';
 import { ImageBlock } from '../../blocks/layout/ImageBlock';
+import { PreviewBlock } from '../../blocks/layout/PreviewBlock';
 import { SectionBlock } from '../../blocks/layout/SectionBlock';
 import { SurfaceRenderer } from '../SurfaceRenderer';
 
@@ -10,14 +11,15 @@ type MessageSurfaceLayoutBlock =
   | ContextBlock
   | DividerBlock
   | ImageBlock
-  | SectionBlock;
+  | SectionBlock
+  | PreviewBlock;
 
 export abstract class UiKitParserMessage<OutputElement> extends SurfaceRenderer<
   OutputElement,
   MessageSurfaceLayoutBlock
 > {
   public constructor() {
-    super(['actions', 'context', 'divider', 'image', 'section']);
+    super(['actions', 'context', 'divider', 'image', 'section', 'preview']);
   }
 }