+++ tools/resolve-loki/bin/resolve-loki-references.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+const { resolveLokiReferences } = require('../dist/resolve-loki-references.js');
+
+resolveLokiReferences();