+++ packages/logo/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.31.0](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.30.1...v0.31.0) (2021-12-28)
+
+### Features
+
+- New hooks for element size tracking ([#413](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/413)) ([8ca682c](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/8ca682c636d2e4813f7d346cb881513382be63cf))
+- **onboarding-ui:** add new background layer with new brand redesign ([#554](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/554)) ([6df6e74](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/6df6e74045183d59deee6db73ae19e59ae1a1482))
+
 # [0.30.0](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.29.0...v0.30.0) (2021-10-06)
 
 ### Features